Now will not overwrite files when not --force, reformate code

+++ src/CrudGeneratorFileCreator.php
@@ -8,31 +8,37 @@ use Illuminate\Support\Str;
 use DB;
 use Artisan;
 
-class CrudGeneratorFileCreator 
+class CrudGeneratorFileCreator
 {
     public $templateName = '';
     public $path = '';
     public $options = [];
     public $deletePrevious = false;
     public $output = null;
-   
 
- 
+
     public function __construct()
     {
 
     }
 
-  
-    public function Generate() {
-        $c = $this->renderWithData($this->customTemplateOfDefault($this->templateName), $this->options);
-        file_put_contents($this->path, $c);
-        $this->output->info('Created Controller: '.$this->path);
+
+    public function Generate()
+    {
+
+        if (!$this->options['force'] && file_exists($this->path)) {
+            $this->output->info('SKIP: ' . $this->path);
+        } else {
+            $c = $this->renderWithData($this->customTemplateOfDefault($this->templateName), $this->options);
+            file_put_contents($this->path, $c);
+            $this->output->info('Created: ' . $this->path);
+        }
+
 
     }
 
-
-    protected function renderWithData($template_path, $data) {
+    protected function renderWithData($template_path, $data)
+    {
         $template = file_get_contents($template_path);
         $template = $this->renderForeachs($template, $data);
         $template = $this->renderIFs($template, $data);
@@ -40,9 +46,10 @@ class CrudGeneratorFileCreator
         return $template;
     }
 
-    protected function renderVariables($template, $data) {
-        $callback = function ($matches) use($data) {
-            if(array_key_exists($matches[1], $data)) {
+    protected function renderVariables($template, $data)
+    {
+        $callback = function ($matches) use ($data) {
+            if (array_key_exists($matches[1], $data)) {
                 return $data[$matches[1]];
             }
             return $matches[0];
@@ -51,23 +58,23 @@ class CrudGeneratorFileCreator
         return $template;
     }
 
-    protected function renderForeachs($template, $data) {
-        $callback = function ($matches) use($data) {
+    protected function renderForeachs($template, $data)
+    {
+        $callback = function ($matches) use ($data) {
             $rep = $matches[0];
             $rep = preg_replace('/\[\[\s*foreach:\s*(.+?)\s*\]\](\r?\n)?/s', '', $rep);
             $rep = preg_replace('/\[\[\s*endforeach\s*\]\](\r?\n)?/s', '', $rep);
             $ret = '';
-            if(array_key_exists($matches[1], $data) && is_array($data[$matches[1]])) {
+            if (array_key_exists($matches[1], $data) && is_array($data[$matches[1]])) {
 
                 $parent = $data[$matches[1]];
                 foreach ($parent as $i) {
                     $d = [];
-                    if(is_array($i)) {
+                    if (is_array($i)) {
                         foreach ($i as $key => $value) {
-                            $d['i.'.$key] = $value;
+                            $d['i.' . $key] = $value;
                         }
-                    }
-                    else {
+                    } else {
                         $d['i'] = $i;
                     }
                     $rep2 = $this->renderIFs($rep, array_merge($d, $data));
@@ -75,53 +82,55 @@ class CrudGeneratorFileCreator
                     $ret .= $rep2;
                 }
                 return $ret;
+            } else {
+                return $mat;
             }
-            else {
-                return $mat;    
-            }
-            
+
         };
         $template = preg_replace_callback('/\[\[\s*foreach:\s*(.+?)\s*\]\](\r?\n)?((?!endforeach).)*\[\[\s*endforeach\s*\]\](\r?\n)?/s', $callback, $template);
         return $template;
     }
 
-    protected function getValFromExpression($exp, $data) {
-        if(Str::contains($exp, "'")) {
-            return trim($exp,"'");    
-        }
-        else {
-            if(array_key_exists($exp, $data)) {
+    protected function getValFromExpression($exp, $data)
+    {
+        if (Str::contains($exp, "'")) {
+            return trim($exp, "'");
+        } else {
+            if (array_key_exists($exp, $data)) {
                 return $data[$exp];
-            }
-            else return null;
+            } else return null;
         }
     }
 
-    protected function renderIFs($template, $data) {
-        $callback = function ($matches) use($data) {
+    protected function renderIFs($template, $data)
+    {
+        $callback = function ($matches) use ($data) {
             $rep = $matches[0];
             $rep = preg_replace('/\[\[\s*if:\s*(.+?)\s*([!=]=)\s*(.+?)\s*\]\](\r?\n)?/s', '', $rep);
             $rep = preg_replace('/\[\[\s*endif\s*\]\](\r?\n)?/s', '', $rep);
             $ret = '';
             $val1 = $this->getValFromExpression($matches[1], $data);
             $val2 = $this->getValFromExpression($matches[3], $data);
-            if($matches[2] == '==' && $val1 == $val2) { $ret .= $rep; }
-            if($matches[2] == '!=' && $val1 != $val2) { $ret .= $rep; }
-            
+            if ($matches[2] == '==' && $val1 == $val2) {
+                $ret .= $rep;
+            }
+            if ($matches[2] == '!=' && $val1 != $val2) {
+                $ret .= $rep;
+            }
+
             return $ret;
         };
         $template = preg_replace_callback('/\[\[\s*if:\s*(.+?)\s*([!=]=)\s*(.+?)\s*\]\](\r?\n)?((?!endif).)*\[\[\s*endif\s*\]\](\r?\n)?/s', $callback, $template);
         return $template;
     }
 
-   
+    protected function customTemplateOfDefault($template_name)
+    {
+        $trypath = base_path() . '/resources/templates/' . $template_name . '.tpl.php';
 
-    protected function customTemplateOfDefault($template_name) {
-        $trypath = base_path().'/resources/templates/'.$template_name.'.tpl.php';
 
-
-        if(file_exists($trypath)) return $trypath;
-        return __DIR__.'/Templates/'.$template_name.'.tpl.php';
+        if (file_exists($trypath)) return $trypath;
+        return __DIR__ . '/Templates/' . $template_name . '.tpl.php';
     }
 
 }
